SplitLayout redesign.
Previous SplitLayout were not working correctly.  The members were not
sized correctly and could also cause some overflow problems.  The new
implementations fixes those problems.  The only problem left is to find
how to block resizing of the top/left most component to a negative
value, which causes the whole layout to grow in size.